Module 1 project
This is a project that tests my knowledge of all the things I have learned so far. This includes data types, for-loops, if statements, operators, and more. So far, I am confident in the coding skills I have learned in module one. The prompt to this project was this: "The moon’s gravity is about 17 percent that of Earth’s. Write a program that displays a table that shows Earth pounds and their equivalent moon weight. Have the table run from 1 to 100 pounds, output a newline every 25 pounds."