Crtical Skill 1.7: Reading Inputs
In this lesson, I learned how to use the std::cin object to accept user input. This allowed me to automate the process of inserting a new int for length and width to compute the area from the previous problem. I also optimized some of my code further and made it tidier, such as declaring my variables in one line and eliminating the need to create a separate variable for area.