Critical Skill 1.9: Two Control Statements (for)
In this program, I learned how to implement a for loop through the template of for(initialization; condition; increment). I also learned that instead of using "var = var + 1" for the increment, we can just use "var++" because the ++ increases the operand by 1. My program prints out 100 numbers, with every number creating a new line until 100, because the condition was "<= 100."