CRITICAL SKILL 2.3: A Closer Look at Variables
In this lesson, I learned how to initialize variables, initialize variables with normally declared variables, and dynamically initialize variables. In this program, I implemented all the things I learned and created a calculator that computes the volume of a sphere with a radius of 10. Although the code was pretty short, it helped me understand how to make code easier to read and some instances where initializing variables would be helpful.